ran ruff

+++ python/mlc_llm/model/baichuan/baichuan_loader.py
@@ -13,7 +13,9 @@ from mlc_llm.quantization import Quantization
 from .baichuan_model import BaichuanConfig, BaichuanForCausalLM
 
 
-def huggingface(model_config: BaichuanConfig, quantization: Quantization) -> ExternMapping:
+def huggingface(
+    model_config: BaichuanConfig, quantization: Quantization
+) -> ExternMapping:
     """Returns a parameter mapping that maps from the names of MLC LLM parameters to
     the names of HuggingFace PyTorch parameters.
 
@@ -52,7 +54,9 @@ def huggingface(model_config: BaichuanConfig, quantization: Quantization) -> Ext
                 f"{mlp}.up_proj.weight",
             ],
             functools.partial(
-                lambda gate, up, dtype: np.concatenate([gate, up], axis=0).astype(dtype),
+                lambda gate, up, dtype: np.concatenate([gate, up], axis=0).astype(
+                    dtype
+                ),
                 dtype=mlc_param.dtype,
             ),
         )
