+++ python/mlc_llm/model/eagle/eagle_loader.py
@@ -70,7 +70,9 @@ def huggingface(model_config: EagleConfig, quantization: Quantization) -> Extern
                 f"{mlp}.up_proj.weight",
             ],
             functools.partial(
-                lambda gate, up, dtype: np.concatenate([gate, up], axis=0).astype(dtype),
+                lambda gate, up, dtype: np.concatenate([gate, up], axis=0).astype(
+                    dtype
+                ),
                 dtype=mlc_param.dtype,
             ),
         )
@@ -167,6 +169,8 @@ def awq(model_config: EagleConfig, quantization: Quantization) -> ExternMapping:
             mapping.add_mapping(
                 mlc_name,
                 [mlc_name],
-                functools.partial(lambda x, dtype: x.astype(dtype), dtype=mlc_param.dtype),
+                functools.partial(
+                    lambda x, dtype: x.astype(dtype), dtype=mlc_param.dtype
+                ),
             )
     return mapping
