+++ python/mlc_llm/model/gemma2/gemma2_loader.py
@@ -13,7 +13,9 @@ from mlc_llm.quantization import Quantization
 from .gemma2_model import Gemma2Config, Gemma2ForCausalLM
 
 
-def huggingface(model_config: Gemma2Config, quantization: Quantization) -> ExternMapping:
+def huggingface(
+    model_config: Gemma2Config, quantization: Quantization
+) -> ExternMapping:
     """Returns a parameter mapping that maps from the names of MLC LLM parameters to
     the names of HuggingFace PyTorch parameters.
 
@@ -69,7 +71,9 @@ def huggingface(model_config: Gemma2Config, quantization: Quantization) -> Exter
                 f"{mlp}.up_proj.weight",
             ],
             functools.partial(
-                lambda gate, up, dtype: np.concatenate([gate, up], axis=0).astype(dtype),
+                lambda gate, up, dtype: np.concatenate([gate, up], axis=0).astype(
+                    dtype
+                ),
                 dtype=mlc_param.dtype,
             ),
         )
