+++ python/mlc_llm/model/minicpm/minicpm_loader.py
@@ -13,7 +13,9 @@ from mlc_llm.quantization import Quantization
 from .minicpm_model import MiniCPMConfig, MiniCPMForCausalLM
 
 
-def huggingface(model_config: MiniCPMConfig, quantization: Quantization) -> ExternMapping:
+def huggingface(
+    model_config: MiniCPMConfig, quantization: Quantization
+) -> ExternMapping:
     """Returns a parameter mapping that maps from the names of MLC LLM parameters to
     the names of HuggingFace PyTorch parameters.
 
@@ -55,7 +57,9 @@ def huggingface(model_config: MiniCPMConfig, quantization: Quantization) -> Exte
                     f"{attn}.v_proj.{weight_type}",
                 ],
                 functools.partial(
-                    lambda q, k, v, dtype: np.concatenate([q, k, v], axis=0).astype(dtype),
+                    lambda q, k, v, dtype: np.concatenate([q, k, v], axis=0).astype(
+                        dtype
+                    ),
                     dtype=mlc_param.dtype,
                 ),
             )
@@ -73,7 +77,9 @@ def huggingface(model_config: MiniCPMConfig, quantization: Quantization) -> Exte
                     f"{mlp}.up_proj.weight",
                 ],
                 functools.partial(
-                    lambda gate, up, dtype: np.concatenate([gate, up], axis=0).astype(dtype),
+                    lambda gate, up, dtype: np.concatenate([gate, up], axis=0).astype(
+                        dtype
+                    ),
                     dtype=mlc_param.dtype,
                 ),
             )
@@ -88,7 +94,9 @@ def huggingface(model_config: MiniCPMConfig, quantization: Quantization) -> Exte
             def combine_expert_gate_up(*hf_params, dtype):
                 stack = []
                 for i in range(0, len(hf_params), 2):
-                    stack.append(np.concatenate([hf_params[i], hf_params[i + 1]], axis=0))
+                    stack.append(
+                        np.concatenate([hf_params[i], hf_params[i + 1]], axis=0)
+                    )
                 return np.stack(stack, axis=0).astype(dtype)
 
             mapping.add_mapping(
